+++ crates/rslint_parser/src/syntax/program.rs
@@ -6,10 +6,9 @@ use super::stmt::{parse_statements, semi, variable_declaration_statement};
 use super::typescript::*;
 use crate::parser::ParserProgress;
 use crate::syntax::class::parse_class_declaration;
-use crate::syntax::function::parse_function_declaration;
+use crate::syntax::function::{is_at_async_function, parse_function_declaration, LineBreak};
 use crate::syntax::object::parse_object_expression;
 use crate::syntax::stmt::directives;
-use crate::syntax::util::{is_at_async_function, LineBreak};
 use crate::ParsedSyntax::Present;
 use crate::{SyntaxKind::*, *};
 use syntax::stmt::FOLLOWS_LET;
